New features        : - CHANGED: Definition of ccubType changed to I32. This allows to better/cleaner use this flag as bitfield in future. Problems solved     : - FIXED: A couple of I32/I64 problems. Bugzilla-Bug-ID     : Additional comments :
git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@5363 4b12f95d-9313-0410-afc9-bd26427041f1

+++ applications/anisotropicdiffusion/src/anisotropicdiffusion_callback.f90
@@ -611,7 +611,7 @@ contains
     type(t_vectorScalar), intent(INOUT)  :: rsolution
 
     ! The type of element used for the FE solution
-    integer(I32), intent(IN) :: ieltype
+    integer, intent(IN) :: ieltype
 
     ! The type of error estimator
     integer, intent(IN) :: ierrorestimator
